requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
openai
pydantic
langserve
